Sending data back to client
In www: Got rid of RequestEmbeddedMap since it was inside GetMapObject anyways.
GetMapObject Sends a map object back to the client via socket.io. No logic yet, i'm just sending the first map object in the array.

In index.html: I refactored the socket code so it sits inside a $(function()... inside the function that sent data to the server and much to my suprise this worked. It receives the stringified map object from the server...which prints an error in the iframe. but hey, we're manipulating the iframe! It should be simple enough to whip this string into a map object and make the embedded map. But I am too tired to do so.