Changed My account button, current ride, Go button
Made changes as discussed on Wednesday. Logout button is now within  My Account screen.
Current ride screen will now update when you create a new ride. (this is just piggybacking on the way Alex has done it)
Added a go button for when we implement filtering the rides. For some reason it doesn't want to go onto the other side of the input bar. Will look into it.