Insert correctly into the DB
Got rid of there being 2 divs with the ID"testMap". ONny the one in Current Ride is.
Passing user_ID to socket so it can store it in Db.
Deleted embedded maps.txt so we don't blow through our quota.
Resuced the size of the loading gif so it fits in the iframe