Notification sent on login
Server keeps an array of which users haven't read their notifications. If they are in the array, we send them a notification even on Register.
If they get a request and are already in the array. We won't add them again.

Also fixed a bug with created_on being a blank string